:recycle: Rewrite Tests
